Refactor: Move converter and key mapping utilities to a new utils module for better organization
