+++ server/computer_use/handlers/utils/converter_utils.py
@@ -10,75 +10,6 @@ from openai.types.chat import (
 
 from server.computer_use.tools.base import BaseAnthropicTool
 
-# Key normalization mappings for computer tools
-KEY_ALIASES = {
-    'Escape': {'esc', 'escape'},
-    'Return': {'enter', 'return'},
-    'Super_L': {'win', 'windows', 'super', 'meta', 'cmd', 'super_l', 'super_r'},
-    'BackSpace': {'backspace'},
-    'Delete': {'del', 'delete'},
-    'Tab': {'tab'},
-    'space': {'space'},
-    'Page_Up': {'pageup'},
-    'Page_Down': {'pagedown'},
-    'Home': {'home'},
-    'End': {'end'},
-    'Up': {'up'},
-    'Down': {'down'},
-    'Left': {'left'},
-    'Right': {'right'},
-    'Print': {'printscreen', 'prtsc'},
-    'ctrl': {'ctrl', 'control', 'ctrl_l', 'ctrl_r'},
-    'shift': {'shift', 'shift_l', 'shift_r'},
-    'alt': {'alt', 'alt_l', 'alt_r', 'option'},
-}
-
-
-def normalize_key_part(part: str) -> str:
-    """
-    Normalize a single key part.
-
-    Args:
-        part: Single key part to normalize
-
-    Returns:
-        Normalized key string
-    """
-    low = part.lower()
-
-    # Check key aliases - find canonical form for any alias
-    for canonical, aliases in KEY_ALIASES.items():
-        if low in aliases:
-            return canonical
-
-    # Function keys
-    if low.startswith('f') and low[1:].isdigit():
-        return f'F{int(low[1:])}'
-
-    # Single letters or digits: keep as-is
-    if len(part) == 1:
-        return part
-
-    return part
-
-
-def normalize_key_combo(combo: str) -> str:
-    """
-    Normalize key combinations for xdotool compatibility.
-
-    Args:
-        combo: Key combination string (e.g., 'ctrl+c', 'alt+tab')
-
-    Returns:
-        Normalized key combination string
-    """
-    if not isinstance(combo, str):
-        return combo
-
-    parts = [p.strip() for p in combo.replace(' ', '').split('+') if p.strip()]
-    normalized = [normalize_key_part(p) for p in parts]
-    return '+'.join(normalized)
-
 
 def _spec_to_openai_chat_function(spec: dict[str, Any]) -> ChatCompletionToolParam:
     name = str(spec.get('name') or '')
